Added new setting for including root domain entries
There was a need for the ability when looking up a entry to also include other items with the same root domain.
For example. oauth.foo.com and sso.foo.com should pull up the same keys. So the best way I thought was to add a checkbox in the settings where you can include all root domain entries. So looking up oauth.foo.com will return keys with foo.com and everything underneth the domain foo.com. I did set this as default disabled to not change behavior of current installations.